lift openmm cif reader to library
